[Desktop] Change ThreadID to long
Summary: Change ThreadID to long

Test Plan:
Messages should send and receive as before. Additionally, if your device has assigned extremely large ThreadIDs, the SMS plugin should no longer crash.

This patch corresponds to the Android-side revision D17517

Reviewers: #kde_connect, nicolasfella

Reviewed By: #kde_connect, nicolasfella

Subscribers: nicolasfella, kdeconnect

Tags: #kde_connect

Differential Revision: https://phabricator.kde.org/D17516

+++ interfaces/conversationmessage.h
@@ -38,7 +38,7 @@ class KDECONNECTINTERFACES_EXPORT ConversationMessage
     Q_PROPERTY(qint64 date READ date)
     Q_PROPERTY(qint32 type READ type)
     Q_PROPERTY(qint32 read READ read)
-    Q_PROPERTY(qint32 threadID READ threadID)
+    Q_PROPERTY(qint64 threadID READ threadID)
     Q_PROPERTY(qint32 uID READ uID)
 
 public:
@@ -73,7 +73,7 @@ public:
 
     ConversationMessage(const qint32& eventField, const QString& body, const QString& address,
                         const qint64& date, const qint32& type, const qint32& read,
-                        const qint32& threadID, const qint32& uID,
+                        const qint64& threadID, const qint32& uID,
                         QObject* parent = Q_NULLPTR);
 
     ConversationMessage(const ConversationMessage& other, QObject* parent = Q_NULLPTR);
@@ -87,7 +87,7 @@ public:
     qint64 date() const { return m_date; }
     qint32 type() const { return m_type; }
     qint32 read() const { return m_read; }
-    qint32 threadID() const { return m_threadID; }
+    qint64 threadID() const { return m_threadID; }
     qint32 uID() const { return m_uID; }
 
     QVariantMap toVariant() const;
@@ -129,7 +129,7 @@ protected:
     /**
      * Tag which binds individual messages into a thread
      */
-    qint32 m_threadID;
+    qint64 m_threadID;
 
     /**
      * Value which uniquely identifies a message
